requirements.txt
requests
tqdm
gdown
Pillow
PyYAML
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
